since strtof does not exist in c89, float numbers are replaced with double

+++ parson.h
@@ -118,7 +118,7 @@ size_t             json_object_get_string_len(const JSON_Object *object, const c
 JSON_Object *      json_object_get_object (const JSON_Object *object, const char *name);
 JSON_Array  *      json_object_get_array  (const JSON_Object *object, const char *name);
 long               json_object_get_number_long(const JSON_Object *object, const char *name); /* returns 0 on fail */
-float              json_object_get_number_float(const JSON_Object *object, const char *name); /* returns 0 on fail */
+double              json_object_get_number_float(const JSON_Object *object, const char *name); /* returns 0 on fail */
 double             json_object_get_number_double(const JSON_Object *object, const char *name); /* returns 0 on fail */
 int                json_object_get_boolean(const JSON_Object *object, const char *name); /* returns -1 on fail */
 
@@ -132,7 +132,7 @@ size_t             json_object_dotget_string_len(const JSON_Object *object, cons
 JSON_Object *      json_object_dotget_object (const JSON_Object *object, const char *name);
 JSON_Array  *      json_object_dotget_array  (const JSON_Object *object, const char *name);
 long               json_object_dotget_number_long(const JSON_Object *object, const char *name); /* returns 0 on fail */
-float              json_object_dotget_number_float(const JSON_Object *object, const char *name); /* returns 0 on fail */
+double             json_object_dotget_number_float(const JSON_Object *object, const char *name); /* returns 0 on fail */
 double             json_object_dotget_number_double(const JSON_Object *object, const char *name); /* returns 0 on fail */
 int                json_object_dotget_boolean(const JSON_Object *object, const char *name); /* returns -1 on fail */
 
@@ -156,7 +156,7 @@ JSON_Status json_object_set_value(JSON_Object *object, const char *name, JSON_Va
 JSON_Status json_object_set_string(JSON_Object *object, const char *name, const char *string);
 JSON_Status json_object_set_string_with_len(JSON_Object *object, const char *name, const char *string, size_t len);  /* length shouldn't include last null character */
 JSON_Status json_object_set_number_long(JSON_Object *object, const char *name, long number);
-JSON_Status json_object_set_number_float(JSON_Object *object, const char *name, float number);
+JSON_Status json_object_set_number_float(JSON_Object *object, const char *name, double number);
 JSON_Status json_object_set_number_double(JSON_Object *object, const char *name, double number);
 JSON_Status json_object_set_boolean(JSON_Object *object, const char *name, int boolean);
 JSON_Status json_object_set_null(JSON_Object *object, const char *name);
@@ -167,7 +167,7 @@ JSON_Status json_object_dotset_value(JSON_Object *object, const char *name, JSON
 JSON_Status json_object_dotset_string(JSON_Object *object, const char *name, const char *string);
 JSON_Status json_object_dotset_string_with_len(JSON_Object *object, const char *name, const char *string, size_t len); /* length shouldn't include last null character */
 JSON_Status json_object_dotset_number_long(JSON_Object *object, const char *name, long number);
-JSON_Status json_object_dotset_number_float(JSON_Object *object, const char *name, float number);
+JSON_Status json_object_dotset_number_float(JSON_Object *object, const char *name, double number);
 JSON_Status json_object_dotset_number_double(JSON_Object *object, const char *name, double number);
 JSON_Status json_object_dotset_boolean(JSON_Object *object, const char *name, int boolean);
 JSON_Status json_object_dotset_null(JSON_Object *object, const char *name);
@@ -190,7 +190,7 @@ size_t        json_array_get_string_len(const JSON_Array *array, size_t index); 
 JSON_Object * json_array_get_object (const JSON_Array *array, size_t index);
 JSON_Array  * json_array_get_array  (const JSON_Array *array, size_t index);
 long          json_array_get_number_long(const JSON_Array *array, size_t index); /* returns 0 on fail */
-float         json_array_get_number_float(const JSON_Array *array, size_t index); /* returns 0 on fail */
+double        json_array_get_number_float(const JSON_Array *array, size_t index); /* returns 0 on fail */
 double        json_array_get_number_double (const JSON_Array *array, size_t index); /* returns 0 on fail */
 int           json_array_get_boolean(const JSON_Array *array, size_t index); /* returns -1 on fail */
 size_t        json_array_get_count  (const JSON_Array *array);
@@ -207,7 +207,7 @@ JSON_Status json_array_replace_value(JSON_Array *array, size_t i, JSON_Value *va
 JSON_Status json_array_replace_string(JSON_Array *array, size_t i, const char* string);
 JSON_Status json_array_replace_string_with_len(JSON_Array *array, size_t i, const char *string, size_t len); /* length shouldn't include last null character */
 JSON_Status json_array_replace_number_long(JSON_Array *array, size_t i, long number);
-JSON_Status json_array_replace_number_float(JSON_Array *array, size_t i, float number);
+JSON_Status json_array_replace_number_float(JSON_Array *array, size_t i, double number);
 JSON_Status json_array_replace_number_double(JSON_Array *array, size_t i, double number);
 JSON_Status json_array_replace_boolean(JSON_Array *array, size_t i, int boolean);
 JSON_Status json_array_replace_null(JSON_Array *array, size_t i);
@@ -221,7 +221,7 @@ JSON_Status json_array_append_value(JSON_Array *array, JSON_Value *value);
 JSON_Status json_array_append_string(JSON_Array *array, const char *string);
 JSON_Status json_array_append_string_with_len(JSON_Array *array, const char *string, size_t len); /* length shouldn't include last null character */
 JSON_Status json_array_append_number_long(JSON_Array *array, long number);
-JSON_Status json_array_append_number_float(JSON_Array *array, float number);
+JSON_Status json_array_append_number_float(JSON_Array *array, double number);
 JSON_Status json_array_append_number_double(JSON_Array *array, double number);
 JSON_Status json_array_append_boolean(JSON_Array *array, int boolean);
 JSON_Status json_array_append_null(JSON_Array *array);
@@ -234,7 +234,7 @@ JSON_Value * json_value_init_array  (void);
 JSON_Value * json_value_init_string (const char *string); /* copies passed string */
 JSON_Value * json_value_init_string_with_len(const char *string, size_t length); /* copies passed string, length shouldn't include last null character */
 JSON_Value * json_value_init_number_long(long number);
-JSON_Value * json_value_init_number_float(float number);
+JSON_Value * json_value_init_number_float(double number);
 JSON_Value * json_value_init_number_double(double number);
 JSON_Value * json_value_init_boolean(int boolean);
 JSON_Value * json_value_init_null   (void);
@@ -247,7 +247,7 @@ JSON_Array  *   json_value_get_array  (const JSON_Value *value);
 const char  *   json_value_get_string (const JSON_Value *value);
 size_t          json_value_get_string_len(const JSON_Value *value); /* doesn't account for last null character */
 long            json_value_get_number_long(const JSON_Value *value);
-float           json_value_get_number_float(const JSON_Value *value);
+double          json_value_get_number_float(const JSON_Value *value);
 double          json_value_get_number_double(const JSON_Value *value);
 int             json_value_get_boolean(const JSON_Value *value);
 JSON_Value  *   json_value_get_parent (const JSON_Value *value);
@@ -259,7 +259,7 @@ JSON_Array  *   json_array  (const JSON_Value *value);
 const char  *   json_string (const JSON_Value *value);
 size_t          json_string_len(const JSON_Value *value); /* doesn't account for last null character */
 long            json_number_long(const JSON_Value *value);
-float           json_number_float(const JSON_Value *value);
+double          json_number_float(const JSON_Value *value);
 double          json_number_double(const JSON_Value *value);
 int             json_boolean(const JSON_Value *value);
